assignment3/requirements.txt
groq
requests
google-genai
cohere
python-dotenv
huggingface_hub